chore(orchestrator): reorganize components into folders
