+++ workspaces/orchestrator/.changeset/late-apples-switch.md
@@ -1,0 +1,21 @@
+---
+'@red-hat-developer-hub/backstage-plugin-orchestrator': patch
+---
+
+### Move components to `ui` folder:
+
+- BaseOrchestratorPage
+- InfoDialog
+- Selector
+- WorkflowInstanceStatusIndicator
+- WorkflowStatus
+
+### Move components to `WorkflowInstancePage` folder:
+
+- Paragraph
+- WorkflowDescriptionModal
+- WorkflowRunDetails
+
+### Move types to `types` folder:
+
+- WorkflowRunDetail
