+++ workspaces/orchestrator/plugins/orchestrator/src/components/WorkflowInstancePage/WorkflowProgress.tsx
@@ -19,7 +19,7 @@ import React from 'react';
 import { ProcessInstanceDTO } from '@red-hat-developer-hub/backstage-plugin-orchestrator-common';
 
 import { compareNodes } from '../../utils/NodeInstanceUtils';
-import { Paragraph } from '../Paragraph';
+import { Paragraph } from './Paragraph';
 import { WorkflowProgressNode } from './WorkflowProgressNode';
 import { fromNodeInstanceToWorkflowProgressNodeModel } from './WorkflowProgressNodeModel';
