+++ workspaces/orchestrator/plugins/orchestrator/src/components/WorkflowInstancePage/WorkflowProgressNode.tsx
@@ -29,7 +29,7 @@ import { ProcessInstanceStatusDTO } from '@red-hat-developer-hub/backstage-plugi
 
 import { VALUE_UNAVAILABLE } from '../../constants';
 import { useWorkflowInstanceStateColors } from '../../hooks/useWorkflowInstanceStatusColors';
-import { Paragraph } from '../Paragraph';
+import { Paragraph } from './Paragraph';
 import { WorkflowProgressNodeModel } from './WorkflowProgressNodeModel';
 
 const WorkflowProgressNodeIcon: React.FC<{
